+++ gem/lib/pagy.rb
@@ -5,12 +5,12 @@ require 'pathname'
 require_relative 'pagy/classes/exceptions'
 require_relative 'pagy/modules/abilities/linkable'
 require_relative 'pagy/modules/abilities/configurable'
-require_relative 'pagy/toolbox/helpers/loader'
+require_relative 'pagy/toolbox/helpers/loaders'
 
 # Top superclass: it defines only what's common to all the subclasses
 # noinspection RubyMismatchedArgumentType
 class Pagy
-  VERSION     = '43.3.1'
+  VERSION     = '43.3.2'
   ROOT        = Pathname.new(__dir__).parent.freeze
   DEFAULT     = { limit: 20, limit_key: 'limit', page_key: 'page' }.freeze
   PAGE_TOKEN  = EscapedValue.new('P ')
@@ -36,7 +36,7 @@ class Pagy
 
   extend Configurable
   include Linkable
-  include Loader
+  include HelperLoader
 
   attr_reader :page, :next, :in, :limit, :options
 
@@ -77,5 +77,10 @@ class Pagy
 
     clean_options = options.delete_if { |k, v| default.key?(k) && (v.nil? || v == '') }
     @options      = default.merge!(clean_options).freeze
+  end
+
+  # Hook module for numeric UI helpers
+  module NumericHelpers
+    include NumericHelperLoader
   end
 end